requirements.txt
flask
opensearch-py 
transformers 
sentence-transformers
python-dotenv